refactor: separate last-checked-at component from main status page component
