+++ src/pages/status/index.tsx
@@ -1,46 +1,17 @@
 import { CheckIcon, WarningIcon } from "@phosphor-icons/react";
-import { useEffect, useState } from "react";
-import { Button } from "@/components/ui/button";
 import { Separator } from "@/components/ui/separator";
 import { Skeleton } from "@/components/ui/skeleton";
 import { useMetadata } from "@/hooks/use-metadata";
 import { useServerStatusQuery } from "@/hooks/use-server-status-query";
+import { LastCheckedAt } from "@/pages/status/components/last-checked-at";
 import { cn } from "@/utils/cn";
 
 export function StatusPage() {
   useMetadata("Página de Status");
-  const [nowTimestamp, setNowTimestamp] = useState<number>(Date.now());
-  const { isSuccess, isError, isLoading, refetch, dataUpdatedAt } =
+  const { isSuccess, isError, isLoading, isRefetching } =
     useServerStatusQuery();
 
-  function handleLastUpdate() {
-    const difference = Math.abs(nowTimestamp - dataUpdatedAt);
-    const validDataUpdatedAt = Number(dataUpdatedAt);
-    if (Number.isNaN(validDataUpdatedAt) || difference < 0)
-      return "data inválida";
-
-    if (difference < 60000) return "agora";
-
-    if (difference < 60 * 60000)
-      return `há ${Math.floor(difference / 60000)}min`;
-
-    if (difference < 24 * 60 * 60000)
-      return `há ${Math.floor(difference / (60 * 60000))}h`;
-
-    return `há ${Math.floor(difference / (24 * 60 * 60000))}d`;
-  }
-
-  function handleRefetch() {
-    setNowTimestamp(Date.now());
-    refetch();
-  }
-
-  useEffect(() => {
-    const intervalId = setInterval(() => {
-      setNowTimestamp(Date.now());
-    }, 60000);
-    return () => clearInterval(intervalId);
-  }, []);
+  const loadingStatus = isLoading || isRefetching;
 
   return (
     <div className="bg-background min-h-screen flex justify-center px-4 py-8 sm:px-6 sm:pt-12">
@@ -51,7 +22,7 @@ export function StatusPage() {
 
         <div className="mt-8 space-y-4 sm:mt-12">
           <div className="bg-surface border border-border rounded-2xl flex flex-col items-start shadow-sm/5 gap-4 p-5 sm:min-h-28 sm:flex-row sm:items-center sm:px-6">
-            {isLoading ? (
+            {loadingStatus ? (
               <Skeleton className="w-12 h-12 bg-foreground-muted/10" />
             ) : (
               <div
@@ -88,7 +59,7 @@ export function StatusPage() {
             )}
             <div className="flex flex-col gap-y-1 min-w-0">
               <h1 className="font-serif font-medium text-xl text-foreground">
-                {isLoading ? (
+                {loadingStatus ? (
                   <Skeleton className="h-6 w-full max-w-full bg-foreground-muted/10 sm:w-96" />
                 ) : (
                   <>
@@ -98,7 +69,7 @@ export function StatusPage() {
                 )}
               </h1>
               <span className="text-foreground-muted">
-                {isLoading ? (
+                {loadingStatus ? (
                   <Skeleton className="h-6 w-full max-w-full bg-foreground-muted/10 sm:w-[31.25rem]" />
                 ) : (
                   <>
@@ -126,7 +97,7 @@ export function StatusPage() {
                   isError && "text-danger",
                 )}
               >
-                {isLoading ? (
+                {loadingStatus ? (
                   <Skeleton className="h-6 w-32 bg-foreground-muted/10" />
                 ) : (
                   <>
@@ -153,7 +124,7 @@ export function StatusPage() {
                   isError && "text-danger",
                 )}
               >
-                {isLoading ? (
+                {loadingStatus ? (
                   <Skeleton className="h-6 w-32 bg-foreground-muted/10" />
                 ) : (
                   <>
@@ -165,24 +136,7 @@ export function StatusPage() {
             </div>
           </div>
 
-          <div className="flex flex-col gap-4 mt-6 sm:flex-row sm:items-center sm:justify-between">
-            <span className="text-foreground-muted text-sm flex items-center gap-x-2">
-              Última alteração:{" "}
-              {isLoading ? (
-                <Skeleton className="w-16 h-6 bg-foreground-muted/10" />
-              ) : (
-                handleLastUpdate()
-              )}
-            </span>
-            <Button
-              disabled={isLoading}
-              onClick={handleRefetch}
-              variant={"outline"}
-              className={"w-full font-semibold hover:bg-black/5 sm:w-48"}
-            >
-              Verificar novamente
-            </Button>
-          </div>
+          <LastCheckedAt isLoading={loadingStatus} />
         </div>
 
         <div className="text-foreground-muted flex gap-x-2 text-sm justify-center mt-16 sm:mt-28">
